+++ src/Pages/EditProfilePage/EditAddressPage/EditAddressForm.js
@@ -7,35 +7,15 @@ import { useHistory}  from 'react-router-dom';
 import {createAddress} from "../../../Services/addressRequest"
 import useProtectedPage from "../../../Hooks/useProtectedPage";
 import { goToProfile } from "../../../Routes/coordinator";
-import { Avatar } from "@material-ui/core";
 
 
 
 const EditAddressForm = () => {
-    useProtectedPage();
-    const [profile, setProfile] = useState({});
+    useProtectedPage();    
     const [address, setAddress] = useState({});
     const[form,onChange,clearFields] = useForm({street: "" , number:[], neighbourhood:"", city:"", state:"", complement:""})
     const history = useHistory();
     
-    
-
-    const getProfile = () => {
-    axios
-        .get(
-        `https://us-central1-missao-newton.cloudfunctions.net/rappi4A/profile`,
-        {
-            headers: {
-                auth: window.localStorage.getItem("token"),
-            },
-        })
-        .then((res) => {
-        setProfile(res.data.user);
-        })
-        .catch((err) => {
-            console.log(err);
-        });
-    };
 
     const getAddress = () => {
         axios
@@ -47,26 +27,22 @@ const EditAddressForm = () => {
                 },
             })
             .then((res) => {
-            setProfile(res.data.address);
+            setAddress(res.data.address);
             
             })
             .catch((err) => {
                 console.log(err);
             });
     };
-
     
 
     useEffect(() => {
-        getProfile();
         getAddress();
     }, []);
 
 
-
     const onSubmitForm = (event) => {
-        createAddress(form, clearFields, history, profile)
-    
+        createAddress(form, clearFields, history, address)
     }
 
     
@@ -80,7 +56,7 @@ const EditAddressForm = () => {
                         fullWidth
                         value= {form.street}               
                         onChange={onChange}
-                        label={profile && profile.street}
+                        label={address && address.street}
                         variant={"outlined"}
                         placeholder={"Rua/Av."}
                        
@@ -92,7 +68,7 @@ const EditAddressForm = () => {
                         name={"number"}
                         value={form.number}
                         onChange={onChange}
-                        label={profile && profile.number}
+                        label={address && address.number}
                         variant={"outlined"}
                         placeholder={"Número"}
                         fullWidth   
@@ -104,7 +80,7 @@ const EditAddressForm = () => {
                         name={"complement"}
                         value={form.complement}
                         onChange={onChange}
-                        label={profile && profile.complement}
+                        label={address && address.complement}
                         variant={"outlined"}
                         placeholder={"Apto./Bloco"}
                         fullWidth
@@ -115,7 +91,7 @@ const EditAddressForm = () => {
                         name={"neighbourhood"}
                         value={form.neighbourhood}
                         onChange={onChange}
-                        label={profile && profile.neighbourhood}
+                        label={address && address.neighbourhood}
                         variant={"outlined"}
                         placeholder={"Bairro"}
                         fullWidth
@@ -127,7 +103,7 @@ const EditAddressForm = () => {
                         name={"city"}
                         value={form.city}
                         onChange={onChange}
-                        label={profile && profile.city}
+                        label={address && address.city}
                         variant={"outlined"}
                         placeholder={"Cidade"}
                         fullWidth
@@ -139,7 +115,7 @@ const EditAddressForm = () => {
                         name={"state"}
                         value={form.state}
                         onChange={onChange}
-                        label={profile && profile.state}
+                        label={address && address.state}
                         variant={"outlined"}
                         placeholder={"Estado"}
                         fullWidth
